[LCR-238] refactor: 오늘 입장자 수 조회 기능 리팩토링
