+++ src/main/java/com/lgcns/domain/entrance/externalApi/EntranceController.java
@@ -1,26 +1,26 @@
 package com.lgcns.domain.entrance.externalApi;
 
-import com.lgcns.domain.entrance.dto.request.EntranceCreateRequest;
+import com.lgcns.domain.entrance.dto.response.DailyEntrantCountResponse;
 import com.lgcns.domain.entrance.service.EntranceService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 @RestController
+@RequestMapping("/popups/{popupId}/dashboard/entrants")
 @RequiredArgsConstructor
-@Tag(name = "6. 방문자 입장 기록 API", description = "QR 인식 후 방문자의 입장 정보를 기록하는 API입니다.")
+@Tag(name = "6-1. 입장 API", description = "입장 관련 API 입니다.")
 public class EntranceController {
 
     private final EntranceService entranceService;
 
-    @PostMapping("/entrances")
-    @Operation(summary = "방문자 입장 기록", description = "QR 인식이 끝난 방문자 정보를 받아 방문자에 대한 정보를 기록합니다.")
-    public ResponseEntity<Void> entranceCreate(@Valid @RequestBody EntranceCreateRequest request) {
-        entranceService.createEntrance(request);
-        return ResponseEntity.status(HttpStatus.CREATED).build();
+    @GetMapping
+    @Operation(summary = "일일 입장자 수 조회", description = "일일 입장자 수를 조회합니다.")
+    public DailyEntrantCountResponse dailyEntrantCountFind(@PathVariable Long popupId) {
+        return entranceService.findDailyEntrantCount(popupId);
     }
 }
