+++ src/main/java/com/lgcns/domain/entrance/repository/EntranceRepository.java
@@ -3,4 +3,5 @@ package com.lgcns.domain.entrance.repository;
 import com.lgcns.domain.entrance.domain.Entrance;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EntranceRepository extends JpaRepository<Entrance, Long> {}
+public interface EntranceRepository
+        extends JpaRepository<Entrance, Long>, EntranceRepositoryCustom {}
