+++ src/main/java/com/lgcns/domain/entrance/service/EntranceService.java
@@ -1,7 +1,10 @@
 package com.lgcns.domain.entrance.service;
 
 import com.lgcns.domain.entrance.dto.request.EntranceCreateRequest;
+import com.lgcns.domain.entrance.dto.response.DailyEntrantCountResponse;
 
 public interface EntranceService {
     void createEntrance(EntranceCreateRequest request);
+
+    DailyEntrantCountResponse findDailyEntrantCount(Long popupId);
 }
