--- src/main/java/com/lgcns/domain/entrant/domain/DailyEntrantCount.java
@@ -1,37 +1,0 @@
-package com.lgcns.domain.entrant.domain;
-
-import com.lgcns.global.model.BaseTimeEntity;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class DailyEntrantCount extends BaseTimeEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "daily_entrant_count_id")
-    private Long id;
-
-    private Long popupId;
-
-    private int entrantCount;
-
-    @Builder
-    private DailyEntrantCount(Long popupId, int entrantCount) {
-        this.popupId = popupId;
-        this.entrantCount = entrantCount;
-    }
-
-    public static DailyEntrantCount createDailyEntrantCount(Long popupId, int entrantCount) {
-        return DailyEntrantCount.builder().popupId(popupId).entrantCount(entrantCount).build();
-    }
-}
