--- src/main/java/com/lgcns/domain/entrant/service/DailyEntrantCountService.java
@@ -1,8 +1,0 @@
-package com.lgcns.domain.entrant.service;
-
-import com.lgcns.domain.entrant.dto.response.DailyEntrantCountResponse;
-
-public interface DailyEntrantCountService {
-
-    DailyEntrantCountResponse findDailyEntrantCount(Long popupId);
-}
